Reworking the config file checker/maker
